Add files via upload
Codebook:
A data frame with 55 observations on the following 5 variables.
gpa: GPA from the students in the study
studyweek:number of hours they studied per week to maintain the GPA
sleepnight:number of hours they slept per week 
out: How often they went out to do something besides school
gender: a factor with levels female and male
The data shows the mean of the students: gpa, studyweek, sleep night and time out.

Summary:
The purpose of this study is to show the GPA of students at Duke University. The study looks at the students number of hours of sleep in comparision to there GPA. As well as there the number of hours that they studied in a week and they also looked at the number of hours they were out and doing things other than school. All of this in comparision to how it impacted there GPA. They also looked at their gender and lookking and a females gpa in comparsion to mens.